found error
Error with mobile devices was due to the min and max height of 100vh, it
is not responsive for some reason. I will have to find another way
around the problem of making each Slide/Div section full screen height.
previous layout works ok on mobile but is not what i am trying to
achieve for the desktop.